% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findOutliers.R
\name{findOutliers}
\alias{findOutliers}
\title{this finds the outliers of two covarites in a SE4 object}
\usage{
findOutliers(diff)
}
\arguments{
\item{diff}{object from findPercentDifferenceAssays}
}
\value{
a kallisto Experiment of differences
}
\description{
this finds the outliers of two covarites in a SE4 object
}
