% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findVariation.R
\name{findVariation}
\alias{findVariation}
\title{this finds the percent difference of two covarites in a SE4 object}
\usage{
findVariation(kexpNew, kexpOrigin)
}
\arguments{
\item{kexpNew}{a kallisto experiment from artemis}

\item{kexpOrigin}{a kallisto experiment considered 'origin' to be compare to}
}
\value{
a list with a bunch of difference data
}
\description{
this finds the percent difference of two covarites in a SE4 object
}
