% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Methods.R
\docType{methods}
\name{SetTxDbLiteMetaClass,KallistoExperiment-method}
\alias{SetTxDbLiteMetaClass,KallistoExperiment-method}
\title{Sets meta columns type to match TxDbLite classes}
\usage{
\S4method{SetTxDbLiteMetaClass}{KallistoExperiment}(object)
}
\description{
Sets meta columns type to match TxDbLite classes
}
